Cambios mínimos

+++ rasi_client/rasi_client/views.py
@@ -23,18 +23,36 @@ def conciliacion_bd(request):
     Args:
         request (_type_): _description_
     """
+    # cuenta los documentos
     count = collection.count_documents({})
     print(count)
+
+    # mientras haya documentos por enviar
     while count != 0:
         query = collection.find_one()
-        post_test(query)
+
+        # TODO! Cuidado con ciclo infinito: podría estar mandando localmente
+        if query.get('method') == 'POST':
+            response = post_test(query)
+        elif query.get('method') == 'PUT':
+            response = put_test(query)
+
+        # evaluamos la respuesta
+        # se perdió la conexión en conciliación
+        if response.status_code not in [200, 204]:
+            break
+
+        # se mantiene la conexión en conciliación
         collection.delete_one({"_id":query['_id']})
         count = collection.count_documents({})
-    print(count)
-    settings.UNSYNC_LOCAL_DB = False
 
-    
-    return HttpResponse("OK")
+
+    if (collection.count_documents({}) > 0):
+        settings.UNSYNC_LOCAL_DB = True
+        return HttpResponse(500)
+    else:
+        settings.UNSYNC_LOCAL_DB = False
+        return HttpResponse(200)
 
 
 def is_online(request):
@@ -104,6 +122,7 @@ def heartbeat(request):
         conciliacion_bd(request)
         return HttpResponse(200)
 
+
 @csrf_exempt
 def get_test(request):
     """
@@ -131,6 +150,7 @@ def get_test(request):
 
     return JsonResponse({'user_id': 1, 'external_data': response_data})
 
+
 @csrf_exempt
 def post_test(request):
     """
@@ -147,17 +167,42 @@ def post_test(request):
     if not settings.REMOTE_DB_ONLINE or settings.UNSYNC_LOCAL_DB:
         response_data = f'WARNING: no hay conexión a la aplicación. Los cambios se guardaran localmente'
 
-        collection.insert_one({'path':request.path,'body':request.body,'method':request.method})
+        collection.insert_one({
+            'path': request.path,
+            'body': request.body,
+            'method': "POST"
+        })
 
+    # envía el post real
     else:
+        return post_real(request)
+
+
+def post_real(request):
+    """
+    auhtor: @c4ts0up
+    Hace un POST de ejemplo. Si está conectado al LB, lo manda a la app en nube. Si no, guarda la petición en la base de datos local para realizarla después
+
+    Args:
+        request (_type_): _description_
+    """
+    url = settings.MANEJADOR_HC_URL + '/historia-clinica/create/'
+
+    if not settings.REMOTE_DB_ONLINE:
+        return JsonResponse({'user_id': 1, 'external_data': 'Error: la aplicación en nube es inaccesible'}, status_code=503)
+    
+    try:
         response = requests.post(url,data=request.body)
 
         if response.status_code == 204:
             response_data = response.text
+            return JsonResponse({'user_id': 1, 'external_data': response_data})
         else:
             response_data = f'Error: {response.status_code}'
+            return JsonResponse({'user_id': 1, 'external_data': 'Error: la aplicación en nube es inaccesible'}, status_code=503)
+    except:
+        return HttpResponse(503)
 
-    return JsonResponse({'user_id': 1, 'external_data': response_data})
 
 @csrf_exempt
 def put_test(request):
@@ -166,3 +211,6 @@ def put_test(request):
     Hace un PUT de ejemplo. Si está conectado al LB, lo manda a la app en nube. Si no, guarda la petición en la base de datos local para realizarla después
     """
     pass
+
+
+def put_real(request):
